+++ changelog.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.8] - 2026-08-25
+## [Unreleased]
+
+### Changed
+
+- Fail targets that declare RGB Matrix without a LED map at import time, explaining that `rgb_matrix.leds` is required, instead of failing compilation with an obscure error
 
 ### Added